Added FeedID column to entries table Feed and Entry IDs converted from long to integer (smart move?) RSSFeed gets in RSSFeed as argument instead of string. Added stub version of RSSFeedManager

+++ Android/TonioTest/TonioTest/src/main/java/com/gnuton/newshub/db/RSSFeedDataSource.java
@@ -76,7 +76,7 @@ public class RSSFeedDataSource extends GenericDataSource {
 
     @Override
     Serializable cursorTo(Cursor cursor) {
-        final long id = cursor.getLong(0);
+        final int id = cursor.getInt(0);
         final String title = cursor.getString(1);
         final String uri = cursor.getString(2);
         Log.d(TAG, "CURSOR " + id + " " + title + " " + uri);
